Adds dynamic plural form determination
Pass xgettext:sourceLang with any of the two-character language codes
listed in PluralForms to explicitly change the plural forms to that
language's plural forms formula based on the formulas provided in the
GNU gettext manual.

If a language is NOT available in the lookup table, or a language is wrong in
the lookup table, pass the formula via the xgettext:rawPluralForm property.
It is probably easier to modify the lookup table than pass a very complex
formula via the property.

The default is "two forms, singular for 1 only", which is the formula used
for English and many other languages worldwide. This default will be used if
neither sourceLang nor rawPluralForm are specified.

Work towards #13