Makes sourceLang optional instead of defaulting to English
@ngocdaothanh requested in
https://github.com/xitrum-framework/scala-xgettext/issues/13#issuecomment-255596466
that the current behavior remain unchanged if no sourceLang option is
provided to the plugin. This accomplishes that.

It also changes the missing language code error behavior to emit the
error to stdout as well as throw an exception. This improves recovering
from the error in Gradle, where the exception message is not propagated
to the user unless the user passes `--stacktrace`. The plugin already
emits a message to stdout upon completion, so this seems to be an
acceptable way to improve the error experience.